requirements.txt
pip>=21.0,<23.2
homeassistant==2023.7.3
pytest-homeassistant-custom-component # Version matching homeassistant
colorlog==6.7.0
pre-commit==3.3.3
black==23.3.0
ruff==0.0.275

statsmodels==0.14.0
matplotlib==3.7.2
pandas-stubs # Version should automatically match pandas version
scikit-learn

# These are duplicated in .pre-commit-config.yaml
reorder-python-imports==3.10.0
mypy==1.4.1
homeassistant-stubs # Version matching homeassistant
voluptuous-stubs==0.1.1
